Fix the way SNMP stuff is scheduled.
To schedule SNMP stuff, snmp_select_info() is called to alter FD sets
to be used with select(). Its last parameter was left unused. In fact,
it should be set to 0 and its value should be checked on return. If
the value is 1, then the provided timer(snmp_timer_wait) should not be
used because it is undefined. Most use of snmp_select_info() set the
timer to NULL in this case. Since we don't want to block, we need to
use our previous value.