Keep current weight on reload when initial weight is not altered.
Weight can be changed by MISC_CHECK when using dynamic option. In case
of reload, the change is lost until the script runs again. We record
the initial weight in a separate variable and use it to check if a
real server has changed instead of using the actual weight.