vrrp: disabled scripts and initially good scripts should be considered as OK.
When a script is not weighted, its failure will lead to a failure of
the associated VRRP instance. However, disabled script and scripts
that are initially good (after a reload) should be considered as
successful and not make the instance fail.

Moreover, a disabled script should not be used when computing script
weights.