vrrp: more informative message when disabling a script due to use of weights
When using a weight for a tracked script, the script is disabled.
However, the warning message said that the weight was ignored. We
change the message to tell that the script is ignored. Moreover, we
don't change its weight since it can be used in another instance, not
in a SYNC group.