Do not set reload flag in the main process.
The main process does not use the reload flag. When setting the reload
flag in the main process and a child dies, the new child will have the
reload flag for a short time. There seems to be some race condition
here. This "fix" seems to circumvent this race condition.